done some changes, almost done i think?
