+++ .idea/workspace.xml
@@ -5,9 +5,10 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="7448c722-a662-457d-81da-39238fb33e99" name="Changes" comment="">
-      <change beforePath="$PROJECT_DIR$/.idea/CatConsole.iml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/CatConsole.iml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/misc.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/misc.xml" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/cash/__init__.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/cash/cash.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/cash.py" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/cash/fs.py" beforeDir="false" afterPath="$PROJECT_DIR$/cash/fs.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -54,6 +55,11 @@
     &quot;vue.rearranger.settings.migration&quot;: &quot;true&quot;
   }
 }</component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="$PROJECT_DIR$/cash" />
+    </key>
+  </component>
   <component name="RunManager">
     <configuration name="CatConsole" type="PythonConfigurationType" factoryName="Python" nameIsGenerated="true">
       <module name="CatConsole" />
@@ -82,8 +88,8 @@
   <component name="SharedIndexes">
     <attachedChunks>
       <set>
-        <option value="bundled-js-predefined-d6986cc7102b-e768b9ed790e-JavaScript-PY-243.21565.199" />
-        <option value="bundled-python-sdk-cab1f2013843-4ae2d6a61b08-com.jetbrains.pycharm.pro.sharedIndexes.bundled-PY-243.21565.199" />
+        <option value="bundled-js-predefined-d6986cc7102b-deb605915726-JavaScript-PY-243.22562.220" />
+        <option value="bundled-python-sdk-0fc6c617c4bd-9a18a617cbe4-com.jetbrains.pycharm.pro.sharedIndexes.bundled-PY-243.22562.220" />
       </set>
     </attachedChunks>
   </component>
@@ -99,6 +105,7 @@
       <workItem from="1738117551161" duration="6206000" />
       <workItem from="1738365161425" duration="8948000" />
       <workItem from="1738617234451" duration="6081000" />
+      <workItem from="1738868252453" duration="2779000" />
     </task>
     <servers />
   </component>
@@ -119,6 +126,6 @@
   <component name="com.intellij.coverage.CoverageDataManagerImpl">
     <SUITE FILE_PATH="coverage/CatConsole$test.coverage" NAME="test Coverage Results" MODIFIED="1738123257043" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="false" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$" />
     <SUITE FILE_PATH="coverage/CatConsole$cash.coverage" NAME="cash Coverage Results" MODIFIED="1738120638743" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="false" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$" />
-    <SUITE FILE_PATH="coverage/CatConsole$CatConsole.coverage" NAME="CatConsole Coverage Results" MODIFIED="1738612586927" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="false" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$" />
+    <SUITE FILE_PATH="coverage/CatConsole$CatConsole.coverage" NAME="CatConsole Coverage Results" MODIFIED="1738872050899" SOURCE_PROVIDER="com.intellij.coverage.DefaultCoverageFileProvider" RUNNER="coverage.py" COVERAGE_BY_TEST_ENABLED="false" COVERAGE_TRACING_ENABLED="false" WORKING_DIRECTORY="$PROJECT_DIR$" />
   </component>
 </project>
