+++ cash/__init__.py
@@ -1,0 +1,3 @@
+from .fs import FS, Dir, File, FSIO
+from .cash import cash
+
